+++ helper/testhelpers/teststorage/teststorage_reusable.go
@@ -5,12 +5,12 @@ import (
 	"io/ioutil"
 	"os"
 
-	"github.com/mitchellh/go-testing-interface"
-
 	hclog "github.com/hashicorp/go-hclog"
 	raftlib "github.com/hashicorp/raft"
 	"github.com/hashicorp/vault/physical/raft"
+	"github.com/hashicorp/vault/sdk/physical"
 	"github.com/hashicorp/vault/vault"
+	"github.com/mitchellh/go-testing-interface"
 )
 
 // ReusableStorage is a physical backend that can be re-used across
@@ -59,8 +59,7 @@ func MakeReusableStorage(t testing.T, logger hclog.Logger, bundle *vault.Physica
 		},
 
 		// No-op
-		Cleanup: func(t testing.T, cluster *vault.TestCluster) {
-		},
+		Cleanup: func(t testing.T, cluster *vault.TestCluster) {},
 	}
 
 	cleanup := func() {
@@ -74,10 +73,7 @@ func MakeReusableStorage(t testing.T, logger hclog.Logger, bundle *vault.Physica
 
 // MakeReusableRaftStorage makes a physical raft backend that can be re-used
 // across multiple test clusters in sequence.
-func MakeReusableRaftStorage(
-	t testing.T, logger hclog.Logger, numCores int,
-	addressProvider raftlib.ServerAddressProvider,
-) (ReusableStorage, StorageCleanup) {
+func MakeReusableRaftStorage(t testing.T, logger hclog.Logger, numCores int, addressProvider raftlib.ServerAddressProvider) (ReusableStorage, StorageCleanup) {
 
 	raftDirs := make([]string, numCores)
 	for i := 0; i < numCores; i++ {
@@ -91,7 +87,7 @@ func MakeReusableRaftStorage(
 			conf.DisablePerformanceStandby = true
 			opts.KeepStandbysSealed = true
 			opts.PhysicalFactory = func(t testing.T, coreIdx int, logger hclog.Logger) *vault.PhysicalBackendBundle {
-				return makeReusableRaftBackend(t, coreIdx, logger, raftDirs[coreIdx], addressProvider)
+				return makeReusableRaftBackend(t, coreIdx, logger, raftDirs[coreIdx], addressProvider, false)
 			}
 		},
 
@@ -120,6 +116,49 @@ func CloseRaftStorage(t testing.T, cluster *vault.TestCluster, idx int) {
 	}
 }
 
+func MakeReusableRaftHAStorage(t testing.T, logger hclog.Logger, numCores int, bundle *vault.PhysicalBackendBundle) (ReusableStorage, StorageCleanup) {
+	raftDirs := make([]string, numCores)
+	for i := 0; i < numCores; i++ {
+		raftDirs[i] = makeRaftDir(t)
+	}
+
+	storage := ReusableStorage{
+		Setup: func(conf *vault.CoreConfig, opts *vault.TestClusterOptions) {
+			opts.KeepStandbysSealed = true
+			opts.PhysicalFactory = func(t testing.T, coreIdx int, logger hclog.Logger) *vault.PhysicalBackendBundle {
+				haBundle := makeReusableRaftBackend(t, coreIdx, logger, raftDirs[coreIdx], nil, true)
+
+				return &vault.PhysicalBackendBundle{
+					Backend:   bundle.Backend,
+					HABackend: haBundle.HABackend,
+				}
+			}
+		},
+
+		// Close open files being used by raft.
+		Cleanup: func(t testing.T, cluster *vault.TestCluster) {
+			for _, core := range cluster.Cores {
+				raftStorage := core.UnderlyingHAStorage.(*raft.RaftBackend)
+				if err := raftStorage.Close(); err != nil {
+					t.Fatal(err)
+				}
+			}
+		},
+	}
+
+	cleanup := func() {
+		if bundle.Cleanup != nil {
+			bundle.Cleanup()
+		}
+
+		for _, rd := range raftDirs {
+			os.RemoveAll(rd)
+		}
+	}
+
+	return storage, cleanup
+}
+
 func makeRaftDir(t testing.T) string {
 	raftDir, err := ioutil.TempDir("", "vault-raft-")
 	if err != nil {
@@ -129,10 +168,7 @@ func makeRaftDir(t testing.T) string {
 	return raftDir
 }
 
-func makeReusableRaftBackend(
-	t testing.T, coreIdx int, logger hclog.Logger, raftDir string,
-	addressProvider raftlib.ServerAddressProvider,
-) *vault.PhysicalBackendBundle {
+func makeReusableRaftBackend(t testing.T, coreIdx int, logger hclog.Logger, raftDir string, addressProvider raftlib.ServerAddressProvider, ha bool) *vault.PhysicalBackendBundle {
 
 	nodeID := fmt.Sprintf("core-%d", coreIdx)
 	conf := map[string]string{
@@ -146,9 +182,16 @@ func makeReusableRaftBackend(
 		t.Fatal(err)
 	}
 
-	backend.(*raft.RaftBackend).SetServerAddressProvider(addressProvider)
-
-	return &vault.PhysicalBackendBundle{
-		Backend: backend,
+	if addressProvider != nil {
+		backend.(*raft.RaftBackend).SetServerAddressProvider(addressProvider)
 	}
+
+	bundle := new(vault.PhysicalBackendBundle)
+
+	if ha {
+		bundle.HABackend = backend.(physical.HABackend)
+	} else {
+		bundle.Backend = backend
+	}
+	return bundle
 }